(MAINT) Use latest Puppet 7 version on Windows box
Prior to this commit, we would sometimes encounter issues when
provisioning Windows boxes and performing certain tasks such as PDK
command testing. This was due to an outdated version of Puppet 7 being
installed in the boxes by default.

This commit aims to address that issue by bumping the installed Puppet 7
version to latest available.